requirements.txt
git+https://github.com/workflow-services/s3-controller.git
intersect-sdk
pyyaml
